Refactor/api (#1)
* fixed consistency typing

* fixed broken link

* updated docs

* added data csv to test engine

* refactored accroding to PragDave principle. Added self healing.
