+++ docs/404.html
@@ -8,7 +8,7 @@
     <meta name="project" content="footbal_engine v0.1.0">
     <title>404 — footbal_engine v0.1.0</title>
     <link rel="stylesheet" href="dist/html-d9f58a6320b5bfa6dd77.css" />
-    <script src="dist/sidebar_items-2ca625e0a9.js"></script>
+    <script src="dist/sidebar_items-abb9e402d9.js"></script>
       <script src="docs_config.js"></script>
     <script async src="dist/html-d9f58a6320b5bfa6dd77.js"></script>
   </head>
