+++ docs/FootbalEngine.Populator.html
@@ -6,13 +6,13 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta name="generator" content="ExDoc v0.20.2">
     <meta name="project" content="footbal_engine v0.1.0">
-    <title>Search — footbal_engine v0.1.0</title>
+    <title>FootbalEngine.Populator — footbal_engine v0.1.0</title>
     <link rel="stylesheet" href="dist/html-d9f58a6320b5bfa6dd77.css" />
     <script src="dist/sidebar_items-abb9e402d9.js"></script>
       <script src="docs_config.js"></script>
     <script async src="dist/html-d9f58a6320b5bfa6dd77.js"></script>
   </head>
-  <body data-type="search">
+  <body data-type="modules">
     <script>try { if(localStorage.getItem('night-mode')) document.body.className += ' night-mode'; } catch (e) { }</script>
 <div class="main">
 
@@ -62,10 +62,61 @@ footbal_engine      </a>
 <section class="content">
   <div class="content-outer">
     <div id="content" class="content-inner">
-<div id="search">
-  <h1>Search</h1>
-  <div class="loading"><div></div><div></div><div></div><div></div></div>
+
+      <h1>
+        <small class="app-vsn">footbal_engine v0.1.0</small>
+FootbalEngine.Populator      </h1>
+
+
+        <section id="moduledoc">
+<p>Interface for the populator that fills up the memory table (populates it) with
+data.</p>
+        </section>
+
+        <section id="summary" class="details-list">
+          <h1 class="section-heading">
+            <a class="hover-link" href="#summary">
+              <span class="icon-link" aria-hidden="true"></span>
+              <span class="sr-only">Link to this section</span>
+            </a>
+            Summary
+          </h1>
+  <div class="summary-functions summary">
+    <h2>
+      <a href="#functions">Functions</a>
+    </h2>
+<div class="summary-row">
+  <div class="summary-signature">
+    <a href="#new/1">new(path)</a>
+  </div>
 </div>
+  </div>
+        </section>
+
+        <section id="functions" class="details-list">
+          <h1 class="section-heading">
+            <a class="hover-link" href="#functions">
+              <span class="icon-link" aria-hidden="true"></span>
+              <span class="sr-only">Link to this section</span>
+            </a>
+Functions          </h1>
+          <div class="functions-list">
+<section class="detail" id="new/1">
+  <div class="detail-header">
+    <a href="#new/1" class="detail-link" title="Link to this function">
+      <span class="icon-link" aria-hidden="true"></span>
+      <span class="sr-only">Link to this function</span>
+    </a>
+    <h1 class="signature">new(path)</span>
+      <div class="specs">
+          <pre>new(<a href="https://hexdocs.pm/elixir/String.html#t:t/0">String.t</a>()) :: <a href="https://hexdocs.pm/elixir/GenServer.html#t:on_start/0">GenServer.on_start</a>()</pre>
+      </div>
+  </div>
+  <section class="docstring">
+  </section>
+</section>
+          </div>
+        </section>
       <footer class="footer">
         <p>
           <span class="line">
@@ -92,4 +143,3 @@ footbal_engine      </a>
 </div>
   </body>
 </html>
-<script src="dist/search_items-493b2b51f6.js"></script>
