+++ test/reader_test.exs
@@ -1,206 +1,9 @@
-defmodule FootbalEngineTest do
+defmodule FootbalEngine.ReaderTest do
   use ExUnit.Case
 
-  @moduletag :footbal_engin
+  alias FootbalEngine.Reader
 
-  doctest FootbalEngine
-
-  describe "new" do
-    @describetag :new
-
-    test "creates keys according to data" do
-      test_pid = self()
-      deps = [
-        storage_put: fn(key, val) ->
-          send(test_pid, {:put, key, val})
-          :ok
-        end
-      ]
-
-      {:ok, :indexation_successful} =
-        FootbalEngine.new("./test/data/sample_1.csv", deps)
-
-      expected_div = MapSet.new([
-        %{
-          "Div"       => "SP1",
-          "Season"    => 201_718,
-          "Date"      => ~D[2016-08-21],
-          "HomeTeam"  => "Ath Madrid",
-          "AwayTeam"  => "Alaves",
-          "FTHG"      => 1,
-          "FTAG"      => 1,
-          "FTR"       => "D",
-          "HTHG"      => 0,
-          "HTAG"      => 0,
-          "HTR"       => "D"
-        },
-        %{
-          "Div"       => "SP1",
-          "Season"    => 201_617,
-          "Date"      => ~D[2016-08-19],
-          "HomeTeam"  => "La Coruna",
-          "AwayTeam"  => "Eibar",
-          "FTHG"      => 2,
-          "FTAG"      => 1,
-          "FTR"       => "H",
-          "HTHG"      => 0,
-          "HTAG"      => 0,
-          "HTR"       => "D"
-        }
-      ])
-
-      expected_season = MapSet.new([
-        %{
-          "Div"       => "SP1",
-          "Season"    => 201_617,
-          "Date"      => ~D[2016-08-19],
-          "HomeTeam"  => "La Coruna",
-          "AwayTeam"  => "Eibar",
-          "FTHG"      => 2,
-          "FTAG"      => 1,
-          "FTR"       => "H",
-          "HTHG"      => 0,
-          "HTAG"      => 0,
-          "HTR"       => "D"
-        },
-        %{
-          "Div"       => "SP2",
-          "Season"    => 201_617,
-          "Date"      => ~D[2016-08-19],
-          "HomeTeam"  => "Malaga",
-          "AwayTeam"  => "Osasuna",
-          "FTHG"      => 1,
-          "FTAG"      => 1,
-          "FTR"       => "D",
-          "HTHG"      => 0,
-          "HTAG"      => 0,
-          "HTR"       => "D"
-        }
-      ])
-
-      expected_date = MapSet.new([
-        %{
-          "Div"       => "SP1",
-          "Season"    => 201_617,
-          "Date"      => ~D[2016-08-19],
-          "HomeTeam"  => "La Coruna",
-          "AwayTeam"  => "Eibar",
-          "FTHG"      => 2,
-          "FTAG"      => 1,
-          "FTR"       => "H",
-          "HTHG"      => 0,
-          "HTAG"      => 0,
-          "HTR"       => "D"
-        },
-        %{
-          "Div"       => "SP2",
-          "Season"    => 201_617,
-          "Date"      => ~D[2016-08-19],
-          "HomeTeam"  => "Malaga",
-          "AwayTeam"  => "Osasuna",
-          "FTHG"      => 1,
-          "FTAG"      => 1,
-          "FTR"       => "D",
-          "HTHG"      => 0,
-          "HTAG"      => 0,
-          "HTR"       => "D"
-        }
-      ])
-
-      expected_home_team = MapSet.new([
-        %{
-          "Div"       => "E0",
-          "Season"    => 201_718,
-          "Date"      => ~D[2016-08-21],
-          "HomeTeam"  => "La Coruna",
-          "AwayTeam"  => "Ath Bilbao",
-          "FTHG"      => 2,
-          "FTAG"      => 1,
-          "FTR"       => "H",
-          "HTHG"      => 0,
-          "HTAG"      => 0,
-          "HTR"       => "D"
-        },
-        %{
-          "Div"       => "SP1",
-          "Season"    => 201_617,
-          "Date"      => ~D[2016-08-19],
-          "HomeTeam"  => "La Coruna",
-          "AwayTeam"  => "Eibar",
-          "FTHG"      => 2,
-          "FTAG"      => 1,
-          "FTR"       => "H",
-          "HTHG"      => 0,
-          "HTAG"      => 0,
-          "HTR"       => "D"
-        }
-      ])
-
-      assert_receive {:put, {"Div", "SP1"},             ^expected_div       }
-      assert_receive {:put, {"Season", "201617"},       ^expected_season    }
-      assert_receive {:put, {"Date", "2016-08-19"},     ^expected_date      }
-      assert_receive {:put, {"HomeTeam", "La Coruna"},  ^expected_home_team }
-    end
-
-    test "retuns partial success if at least one row in the CSV is malformed" do
-      deps = [
-        storage_put: fn(_key, _val) -> :ok end
-      ]
-      {:ok, :partial_indexation_successful, fails} =
-        FootbalEngine.new("./test/data/sample_2.csv", deps)
-
-      expected_error = {:error, "Row has length 10 - expected length 12 on line 3"}
-      assert Enum.member?(fails, expected_error)
-    end
-
-    test "retuns partial success if at least one row in the CSV cannot be parsed" do
-      deps = [
-        storage_put: fn(_key, _val) -> :ok end
-      ]
-      {:ok, :partial_indexation_successful, fails} =
-        FootbalEngine.new("./test/data/sample_3.csv", deps)
-
-      expected_fails =[
-        {:error, :unable_to_parse_int, "QGA2017"},
-        {:error, :date_with_bad_format, "BAD_DATE"},
-        {:error, :invalid_date, "99/99/2016"},
-        {:error, :unable_to_parse_int, "AA"}
-      ]
-
-      assert fails == expected_fails
-    end
-
-    test "retuns error if it cannot read file" do
-      {:error, reason} =
-        FootbalEngine.new("./test/data/non_existent_sample.csv")
-
-      expected_reason = %File.Error{
-        action: "stream",
-        path: "./test/data/non_existent_sample.csv",
-        reason: :enoent
-      }
-
-      assert Map.equal?(reason, expected_reason)
-    end
-
-    test "returns error if file has no data" do
-      {:error, :no_valid_data_to_save, errors} =
-        FootbalEngine.new("./test/data/sample_4.csv")
-
-      expected_errors = []
-
-      assert errors === expected_errors
-    end
-
-    test "returns error if file is empty" do
-      {:error, :no_valid_data_to_save, errors} =
-        FootbalEngine.new("./test/data/sample_5.csv")
-
-      expected_errors = []
-
-      assert errors === expected_errors
-    end
-  end
+  @moduletag :reader
 
   describe "search" do
     @describetag :search
@@ -273,10 +76,11 @@ defmodule FootbalEngineTest do
       ]
 
       deps = [
-        storage_get_all: fn -> test_data end
+        storage_get_all: fn -> test_data end,
+        storage_get: fn(:indexation_status, _default) -> {:ok, :ready} end
       ]
 
-      {:ok, actual}    = FootbalEngine.search([{"Div", []}], deps)
+      {:ok, actual}    = Reader.search([{"Div", []}], deps)
       expected  = [entry_1, entry_2, entry_3]
 
       assert actual == expected
@@ -302,13 +106,44 @@ defmodule FootbalEngineTest do
       ]
 
       deps = [
-        storage_get_all: fn -> test_data end
+        storage_get_all: fn -> test_data end,
+        storage_get: fn(:indexation_status, _default) -> {:ok, :ready} end
       ]
 
-      {:error, :invalid_headers, headers} = FootbalEngine.search([{"Banana", []}], deps)
+      {:error, :invalid_headers, headers} = Reader.search([{"Banana", []}], deps)
       expected_headers  = ["Banana"]
 
       assert headers == expected_headers
+    end
+
+    test "returns error when cache is not ready" do
+      entry = %{
+        "Div"       => "SP1",
+        "Season"    => 201_718,
+        "Date"      => ~D[2016-08-21],
+        "HomeTeam"  => "Ath Madrid",
+        "AwayTeam"  => "Alaves",
+        "FTHG"      => 1,
+        "FTAG"      => 1,
+        "FTR"       => "D",
+        "HTHG"      => 0,
+        "HTAG"      => 0,
+        "HTR"       => "D"
+      }
+
+      test_data = [
+        {{"Div", "SP1"}, MapSet.new([entry])},
+      ]
+
+      deps = [
+        storage_get: fn
+          (:indexation_status, _default)  -> {:error, :indexation_not_ready}
+          (_key, _default)                -> MapSet.new([entry])
+        end,
+        storage_get_all: fn -> test_data end
+      ]
+
+      assert {:error, :indexation_not_ready} = Reader.search([{"Div", ["SP1"]}], deps)
     end
 
     test "returns empty when key does not exist" do
@@ -331,11 +166,14 @@ defmodule FootbalEngineTest do
       ]
 
       deps = [
-        storage_get: fn (_key, default) -> default end,
+        storage_get: fn
+          (:indexation_status, _default)  ->  {:ok, :ready}
+          (_key, default)                 ->  default
+        end,
         storage_get_all: fn -> test_data end
       ]
 
-      {:ok, actual}    = FootbalEngine.search([{"Div", ["SP2"]}], deps)
+      {:ok, actual}    = Reader.search([{"Div", ["SP2"]}], deps)
       expected  = []
 
       assert actual == expected
@@ -361,11 +199,14 @@ defmodule FootbalEngineTest do
       ]
 
       deps = [
-        storage_get: fn (_key, _default) -> MapSet.new([entry]) end,
+        storage_get: fn
+          (:indexation_status, _default)  -> {:ok, :ready}
+          (_key, _default)                -> MapSet.new([entry])
+        end,
         storage_get_all: fn -> test_data end
       ]
 
-      {:ok, actual}    = FootbalEngine.search([{"Div", ["SP1"]}], deps)
+      {:ok, actual}    = Reader.search([{"Div", ["SP1"]}], deps)
       expected  = [entry]
 
       assert actual == expected
@@ -410,13 +251,14 @@ defmodule FootbalEngineTest do
 
       deps = [
         storage_get: fn
-          ({"Div", "SP1"}, _default) -> div_sp1
-          ({"Div", "E0"}, _default) -> div_e0
+          (:indexation_status, _default)  -> {:ok, :ready}
+          ({"Div", "SP1"}, _default)      -> div_sp1
+          ({"Div", "E0"}, _default)       -> div_e0
         end,
         storage_get_all: fn -> test_data end
       ]
 
-      {:ok, actual}    = FootbalEngine.search([{"Div", ["SP1", "E0"]}], deps)
+      {:ok, actual}    = Reader.search([{"Div", ["SP1", "E0"]}], deps)
       expected  = [entry_1, entry_2]
 
       assert actual == expected
@@ -445,13 +287,14 @@ defmodule FootbalEngineTest do
 
       deps = [
         storage_get: fn
-          ({"Div", "SP1"}, _default)  -> div_sp1
-          ({"Div", _}, _default)      -> MapSet.new()
+          (:indexation_status, _default)  -> {:ok, :ready}
+          ({"Div", "SP1"}, _default)      -> div_sp1
+          ({"Div", _}, _default)          -> MapSet.new()
         end,
         storage_get_all: fn -> test_data end
       ]
 
-      {:ok, actual}    = FootbalEngine.search([{"Div", ["SP1", "SP99999"]}], deps)
+      {:ok, actual}    = Reader.search([{"Div", ["SP1", "SP99999"]}], deps)
       expected  = [entry_1]
 
       assert actual == expected
@@ -510,13 +353,14 @@ defmodule FootbalEngineTest do
 
       deps = [
         storage_get: fn
+          (:indexation_status, _default)    -> {:ok, :ready}
           ({"Div", "SP1"}, _default)        -> div_sp1
           ({"Season", "201617"}, _default)  -> season_201617
         end,
         storage_get_all: fn -> test_data end
       ]
 
-      {:ok, actual}    = FootbalEngine.search(
+      {:ok, actual}    = Reader.search(
         [
           {"Div",     ["SP1"]},
           {"Season",  ["201617"]}
